chore: improve typing
