+++ src/data/model/PostMeta.ts
@@ -1,0 +1,8 @@
+interface PostMeta {
+  title: string
+  previewImgUri?: string
+  route: string
+  repo?: string
+}
+
+export type { PostMeta }
